feat: req로 받은 값 검증 및 예외처리 추가

+++ controllers/commentController.js
@@ -3,10 +3,12 @@ const {
   InternalServerError,
 } = require('../middleware/customError');
 const { save, update, findByUserId, deleteById } = require('../model/Comment');
+const { getTimestamp } = require('../utils/dayUtil');
 const { postValidator } = require('../utils/validation');
 
 const createComment = async (req, res, next) => {
   try {
+    // 클라이언트 요청 데이터 (내용, 유저 아이디, 게시글 아이디)
     const { content, user_id, post_id } = req.body;
 
     // 유효성 검사
@@ -17,16 +19,17 @@ const createComment = async (req, res, next) => {
     }
 
     if (!user_id || !post_id) {
-      next(new BadRequest('댓글 생성에 필수 정보가 누락되었습니다.'));
+      return next(new BadRequest('댓글 생성에 필수 정보가 누락되었습니다.'));
     }
 
     const commentData = {
       content,
-      created_at: timestamp(),
-      modified_at: timestamp(),
+      created_at: getTimestamp(),
+      modified_at: getTimestamp(),
       user_id,
       post_id,
     };
+
     const newComment = await save(commentData);
 
     const { id } = newComment;
@@ -38,21 +41,15 @@ const createComment = async (req, res, next) => {
       },
     });
   } catch (error) {
-    next(new InternalServerError());
+    return next(new InternalServerError('댓글 작성을 실패했습니다.'));
   }
 };
-//날짜 변환 함수
-const timestamp = () => {
-  const today = new Date();
-  // 미국시간 기준이니까 9를 더해주면 대한민국 시간
-  today.setHours(today.getHours() + 9);
-  // 문자열로 바꿔주고 T를 빈칸으로 바꿔주면 yyyy-mm-dd hh:mm:ss 이런 형식 나옴
-  return today.toISOString().replace('T', ' ').substring(0, 19);
-};
+
 //댓글 수정
 const updateComment = async (req, res, next) => {
   try {
-    const { content, post_id, user_id, comment_id } = req.body;
+    // 클라이언트 요청 데이터 (내용, 게시글 아이디, 유저 아이디, 댓글 아이디)
+    const { content, post_id, user_id, comment_id } = req.body; 
 
     // 유효성 검사
     const commentValidation = postValidator.comment(content);
@@ -61,9 +58,13 @@ const updateComment = async (req, res, next) => {
       return next(new BadRequest(commentValidation.message));
     }
 
+    if(!post_id || !user_id || !comment_id) {
+      return next(new BadRequest('댓글 수정에 필수 정보가 누락되었습니다.'));
+    }
+
     const commentData = {
       content,
-      modified_at: timestamp(),
+      modified_at: getTimestamp(),
       user_id,
       post_id,
       comment_id,
@@ -78,7 +79,7 @@ const updateComment = async (req, res, next) => {
       },
     });
   } catch (error) {
-    next(new InternalServerError());
+    return next(new InternalServerError('댓글 수정에 실패했습니다.'));
   }
 };
 //해당 댓글 조회
@@ -86,18 +87,22 @@ const findCommentUser = async (req, res, next) => {
   try {
     const userId = Number(req.user.id);
 
+    if(!userId) {
+      return next(new BadRequest('댓글 조회시 해당 유저의 정보가 없습니다.'));
+    }
+
     const comments = await findByUserId(userId);
 
     if (!comments) {
-      next(new BadRequest('해당 유저의 댓글을 조회하는데 실패했습니다.'));
+      return next(new BadRequest('해당 유저의 댓글을 조회하는데 실패했습니다.'));
     }
 
     res.status(200).json({
-      message: '댓글 조회를 성공했습니다.',
+      message: '해당 유저의 댓글 조회를 성공했습니다.',
       data: comments,
     });
   } catch (error) {
-    next(new InternalServerError());
+    return next(new InternalServerError('해당 유저의 댓글을 조회하는데 실패했습니다'));
   }
 };
 
@@ -106,10 +111,14 @@ const deleteComment = async (req, res, next) => {
   try {
     const id = Number(req.params.comment_id);
 
+    if(!id) {
+      return next(new BadRequest('댓글 삭제 중 게시글 정보가 누락되었습니다.'));
+    }
+
     const comment = await deleteById(id);
 
     if (!comment) {
-      next(new BadRequest('댓글 삭제에 실패했습니다.'));
+      return next(new BadRequest('댓글 삭제에 실패했습니다.'));
     }
 
     res.status(200).json({
@@ -119,7 +128,7 @@ const deleteComment = async (req, res, next) => {
       },
     });
   } catch (error) {
-    next(new InternalServerError());
+    return next(new InternalServerError('댓글 삭제에 실패했습니다.'));
   }
 };
